# docs/openapi/serviceability-api.yaml
openapi: 3.0.3
info:
  title: Prayog Serviceability API
  description: |
    Comprehensive API for managing service area coverage, geographical entities, location management, and delivery serviceability.

    ## Features
    - **Geographical Management**: Countries, regions, districts, cities, and areas
    - **Location Types & Aliases**: Flexible location classification and aliasing system
    - **Postal Code Management**: Postal code management with location scope validation
    - **Partner Integration**: External partner service integration and validation
    - **Serviceability Check**: Service area coverage and delivery options

    ## Authentication
    All endpoints require valid authentication. Include your API key in the Authorization header:
    ```
    Authorization: Bearer <your-api-key>
    ```

    ## Response Format
    All successful responses follow this format:
    ```json
    {
      "status": "success",
      "data": { ... }
    }
    ```

    Error responses follow this format:
    ```json
    {
      "status": "error", 
      "message": "Error description",
      "code": "error_code"
    }
    ```
  version: 1.0.0
  contact:
    email: api@prayog.com
    name: Prayog API Support
  license:
    name: Proprietary
    url: https://prayog.com/license
servers:
  - url: "https://api.prayog.com"
    description: Production server
  - url: "https://api.staging.prayog.com"
    description: Staging server
  - url: "http://localhost:8080"
...
security:
  - BearerAuth: []
paths:
...
  /serviceability/v1/countries:
    get:
      tags:
        - countries
      summary: List all countries
      description: Retrieve a paginated list of all countries
      operationId: listCountries
      parameters:
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Successfully retrieved countries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountryListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - countries
      summary: Create a new country
      description: Create a new country with validation
      operationId: createCountry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCountryRequest"
      responses:
        "201":
          description: Country created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountryResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
  /serviceability/v1/countries/{id}:
    parameters:
      - $ref: "#/components/parameters/UUIDParam"
    get:
      tags:
        - countries
      summary: Get country by ID
      description: Retrieve a specific country by its UUID
      operationId: getCountryById
      responses:
        "200":
          description: Successfully retrieved country
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags:
        - countries
      summary: Update country
      description: Update an existing country
      operationId: updateCountry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCountryRequest"
      responses:
        "200":
          description: Country updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountryResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags:
        - countries
      summary: Delete country
      description: Soft delete a country (sets is_active to false)
      operationId: deleteCountry
      responses:
        "204":
          description: Country deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /serviceability/v1/countries/code/{code}:
    parameters:
      - name: code
        in: path
        required: true
        description: Country code (e.g., 'US', 'IN')
        schema:
          type: string
          minLength: 2
          maxLength: 3
    get:
      tags:
        - countries
      summary: Get country by code
      description: Retrieve a country by its country code
      operationId: getCountryByCode
      responses:
        "200":
          description: Successfully retrieved country
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # Country Aliases (Nested)
...
  /serviceability/v1/region-types:
    get:
      tags:
        - region-types
      summary: List all region types
      description: Retrieve a paginated list of all region types
      operationId: listRegionTypes
      parameters:
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Successfully retrieved region types
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionTypeListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - region-types
      summary: Create a new region type
      description: Create a new region type with validation
      operationId: createRegionType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRegionTypeRequest"
      responses:
        "201":
          description: Region type created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionTypeResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
  /serviceability/v1/region-types/{code}:
    parameters:
      - name: code
        in: path
        required: true
        description: Region type code
        schema:
          type: string
    get:
      tags:
        - region-types
      summary: Get region type by code
      description: Retrieve a specific region type by its code
      operationId: getRegionTypeByCode
      responses:
        "200":
          description: Successfully retrieved region type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionTypeResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags:
        - region-types
      summary: Update region type
      description: Update an existing region type
      operationId: updateRegionType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRegionTypeRequest"
      responses:
        "200":
          description: Region type updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionTypeResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags:
        - region-types
      summary: Delete region type
      description: Soft delete a region type
      operationId: deleteRegionType
      responses:
        "204":
          description: Region type deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API authentication
  parameters:
    UUIDParam:
      name: id
      in: path
      required: true
      description: Entity UUID
      schema:
        type: string
        format: uuid
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip for pagination
      schema:
        type: integer
        minimum: 0
        default: 0
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  schemas:
...
    CreateCountryRequest:
      type: object
      required:
        - name
        - code
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Country name
        code:
          type: string
          minLength: 2
          maxLength: 3
          description: ISO country code
        is_active:
          type: boolean
          default: true
    UpdateCountryRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        code:
          type: string
          minLength: 2
          maxLength: 3
        is_active:
          type: boolean
    CountryResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        data:
          $ref: "#/components/schemas/Country"
    CountryListResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          properties:
            countries:
              type: array
              items:
                $ref: "#/components/schemas/Country"
            pagination:
              $ref: "#/components/schemas/PaginationInfo"

    # Region Type Schemas
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorResponse"
              - type: object
                properties:
                  code:
                    type: string
                    enum: [validation_error]
          example:
            status: error
            message: "Field 'name' is required"
            code: validation_error
    ConflictError:
      description: Resource already exists or constraint violation
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorResponse"
              - type: object
                properties:
                  code:
                    type: string
                    enum: [conflict_error]
          example:
            status: error
            message: "Country with code 'US' already exists"
            code: conflict_error
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorResponse"
              - type: object
                properties:
                  code:
                    type: string
                    enum: [internal_error]
          example:
            status: error
            message: "An internal error occurred"
            code: internal_error